unit tests: issue-credential
