+++ aries_cloudagent/protocols/issue_credential/v2_0/messages/cred_proposal.py
@@ -112,20 +112,20 @@ class V20CredProposalSchema(AgentMessageSchema):
 
     @validates_schema
     def validate_fields(self, data, **kwargs):
-        """Validate filter per format."""
+        """Validate attachments per format."""
 
-        def get_filter_attach_by_id(attach_id):
-            """Return filter with input attachment identifier."""
-            for f in filters_attach:
+        def get_attach_by_id(attach_id):
+            """Return attachment with input identifier."""
+            for f in attachments:
                 if f.ident == attach_id:
                     return f
-            raise ValidationError(f"No filter matches attach_id {attach_id} in format")
+            raise ValidationError(f"No attachment for attach_id {attach_id} in formats")
 
         formats = data.get("formats") or []
-        filters_attach = data.get("filters_attach") or []
-        if len(formats) != len(filters_attach):
-            raise ValidationError("Formats/filters length mismatch")
+        attachments = data.get("filters_attach") or []
+        if len(formats) != len(attachments):
+            raise ValidationError("Formats/attachments length mismatch")
 
         for fmt in formats:
-            filt_atch = get_filter_attach_by_id(fmt.attach_id)
-            V20CredFormat.Format.get(fmt.format).validate_filter(filt_atch.content)
+            atch = get_attach_by_id(fmt.attach_id)
+            V20CredFormat.Format.get(fmt.format).validate_filter_attach(atch.content)
